roll out the refactor, rewrite and API changes to BMe fully
